+++ mapmatching/geo/haversineDistance.py
@@ -14,7 +14,7 @@ def haversine_matrix(array1, array2, xy=True, unit=Unit.METERS):
     '''
     if xy:
         array1 = array1[:, ::-1]
-        array2 = array2[:,::-1]
+        array2 = array2[:, ::-1]
     
     dist = haversine_vector(np.repeat(array1, len(array2), axis=0), 
                             np.concatenate([array2] * len(array1)),
@@ -24,7 +24,7 @@ def haversine_matrix(array1, array2, xy=True, unit=Unit.METERS):
 
     return matrix
 
-def haversine_vector_xy(array1, array2, unit=Unit.KILOMETERS, comb=False, normalize=False):
+def haversine_vector_xy(array1, array2, unit=Unit.METERS, comb=False, normalize=False):
     # ensure arrays are numpy ndarrays
     if not isinstance(array1, np.ndarray):
         array1 = np.array(array1)
@@ -32,7 +32,7 @@ def haversine_vector_xy(array1, array2, unit=Unit.KILOMETERS, comb=False, normal
         array2 = np.array(array2)
 
     array1 = array1[:, ::-1]
-    array2 = array2[:,::-1]
+    array2 = array2[:, ::-1]
     ans = haversine_vector(array1, array2, unit, comb, normalize)
     
     return ans
@@ -53,7 +53,7 @@ def coords_pair_dist(o, d, xy=True):
 def points_geoseries_2_ndarray(geoms:gpd.GeoSeries):
     return np.concatenate(geoms.apply(lambda x: x.coords[:]).values).reshape(-1, 2)
 
-def cal_points_seq_distance(points:np.ndarray, xy=True):
+def cal_coords_seq_distance(points:np.ndarray, xy=True):
     if xy:
         points = points.copy()
         points = points[:, ::-1]
@@ -64,9 +64,17 @@ def cal_points_seq_distance(points:np.ndarray, xy=True):
 
 def cal_points_geom_seq_distacne(geoms:gpd.GeoSeries):
     coords = points_geoseries_2_ndarray(geoms)
-    dist, total = cal_points_seq_distance(coords, xy=True)
+    dist, total = cal_coords_seq_distance(coords, xy=True)
 
     return dist, coords
+
+def haversine_geoseries(points1, points2, unit=Unit.METERS, comb=False, normalize=False):
+    coords_0 = points_geoseries_2_ndarray(points1)
+    coords_1 = points_geoseries_2_ndarray(points2)
+    dist = haversine_vector_xy(coords_0, coords_1, unit, comb, normalize)
+
+    return dist
+
 
 if __name__ == "__main__":
     matrix = haversine_matrix(traj_points, points_, xy=True)