+++ mapmatching/geo/ops/resample.py
@@ -7,6 +7,7 @@ from ..haversineDistance import cal_points_geom_seq_distacne
 
 
 def resample_point_seq(points, step=2, last=True):
+    # TODO linear referencing
     if points.shape[0] == 1:
         return gpd.GeoDataFrame(points), np.array([points.iloc[0].coords[0]])