+++ mapmatching/graph/__init__.py
@@ -1,1 +1,1 @@
-from .geograph import GeoDigraph
+from .geograph import GeoDigraph, GeoDigraphLL