+++ mapmatching/match/visualization.py
@@ -9,14 +9,9 @@ import geopandas as gpd
 import matplotlib.pyplot as plt
 from shapely.geometry import box
 from ..utils.parallel_helper import parallel_process
+from ..utils.img import merge_np_imgs
 
-try:
-    from tilemap import plot_geodata, add_basemap
-    TILEMAP_FLAG = True
-except:
-    def plot_geodata(data, *args, **kwargs):
-        return data.plot(*args, **kwargs)
-    TILEMAP_FLAG = False
+from ..geo.vis import plot_geodata, TILEMAP_FLAG, add_basemap
 
 def matching_debug_subplot(traj, edges, item, level, src, dst, ax=None, maximun=None, legend=True, factor=4):
     """Plot the matching situation of one pair of od.
@@ -48,11 +43,11 @@ def matching_debug_subplot(traj, edges, item, level, src, dst, ax=None, maximun=
     point_0.plot(ax=ax, marker="*", label=f'O ({src})', zorder=8)
     point_n.plot(ax=ax, marker="s", label=f'D ({dst})', zorder=8)
     font_style = {'zorder': 8, "ha": "center", "va":"bottom", "zorder":9}
-    trans_p_str = f"{item.trans_prob:.2f}"
+    trans_p_str = f"trans: {item.trans_prob:.2f}"
     if 'dir_prob' in item:
-        trans_p_str += f"({item.dist_prob:.2f}, {item.dir_prob:.2f})"
+        trans_p_str += f"(dist: {item.dist_prob:.2f}, dir: {item.dir_prob:.2f})"
     ax.text(point_0.geometry.x, point_0.geometry.y, trans_p_str, **font_style)
-    ax.text(point_n.geometry.x, point_n.geometry.y, f"{item.observ_prob:.2f}", **font_style)
+    ax.text(point_n.geometry.x, point_n.geometry.y, f"oberv: {item.observ_prob:.2f}", **font_style)
 
     # path
     gpd.GeoDataFrame( item ).T.plot(ax=ax, color='red', label='Path')
@@ -89,49 +84,6 @@ def ops_matching_debug_subplot(traj, edges, item, layer_id, src, dst, maximun):
 
     return array
 
-def merge_imgs_np(arrays, n_row, n_col):
-    """
-    Merge a set of tiles into a single array.
-
-    Parameters
-    ---------
-    tiles : list of mercantile.Tile objects
-        The tiles to merge.
-    arrays : list of numpy arrays
-        The corresponding arrays (image pixels) of the tiles. This list
-        has the same length and order as the `tiles` argument.
-
-    Returns
-    -------
-    img : np.ndarray
-        Merged arrays.
-    extent : tuple
-        Bounding box [west, south, east, north] of the returned image
-        in long/lat.
-    """
-    # get indices starting at zero
-    indices = []
-    for r in range(n_row):
-        for c in range(n_col):
-            indices.append((r, c))
-
-    # the shape of individual tile images
-    h, w, d = arrays[0].shape
-    h = max([i.shape[0] for i in arrays])
-    w = max([i.shape[1] for i in arrays])
-
-    # empty merged tiles array to be filled in
-    img = np.ones((h * n_row, w * n_col, d), dtype=np.uint8) * 255
-
-    for ind, arr in zip(indices, arrays):
-        y, x = ind
-        _h, _w, _ = arr.shape
-        ori_x = x * w + (w - _w) // 2
-        ori_y = y * h + (h - _h) // 2
-        img[ori_y : ori_y + _h, ori_x : ori_x + _w, :] = arr
-
-    return img
-
 def debug_gt_level_parallel(net, traj, graph, layer_id, n_jobs=16):
     df_layer = graph.loc[layer_id]
     if df_layer.empty:
@@ -143,7 +95,7 @@ def debug_gt_level_parallel(net, traj, graph, layer_id, n_jobs=16):
     params = ((traj, net.get_edge([src, dst]), df_layer.loc[src].loc[dst], layer_id, src, dst, _max) 
                 for src, dst in df_layer.index.values)
     img_arrs = parallel_process(ops_matching_debug_subplot, params, n_jobs=n_jobs, pbar_switch=True)
-    img_np = merge_imgs_np(img_arrs, n_rows, n_cols)
+    img_np = merge_np_imgs(img_arrs, n_rows, n_cols)
 
     img = Image.fromarray(img_np).convert("RGB")